+++ index.mjs
@@ -1,3 +1,4 @@
 export { abi } from './build/contracts/ERC721Pagination.json';
+export { abi as interfaceIdAbi } from './build/contracts/ERC721PaginationInterfaceId.json';
 export { abi as interfaceAbi } from './build/contracts/IERC721Pagination.json';
 export const ERC721_PAGINATION_INTREFACE_ID = '0xfca7958f';